do not warn on < 5.8.9
Explicitly assigning undef to a %SIG handler is not supported on < 5.8.9.
This results in warnings, and leads among other things to a failure in the
DBIC test suite.

A plain local() call is sufficient to "unset" the handler on all perls.

(as seen in https://github.com/autarch/DateTime-TimeZone/pull/4)